feat: make metrics works
