+++ Makefile
@@ -13,9 +13,9 @@ GIT_DATE := $(shell git log -1 --format=%ci | cut -d ' ' -f 1)
 
 #? shisui: Build shisui
 shisui:
-	go build -ldflags "-X github.com/zen-eth/shisui/internal/version.gitCommit=$(GIT_COMMIT) -X github.com/zen-eth/shisui/internal/version.gitDate=$(GIT_DATE)" ./cmd/shisui/main.go
+	go build -ldflags "-X github.com/zen-eth/shisui/internal/version.gitCommit=$(GIT_COMMIT) -X github.com/zen-eth/shisui/internal/version.gitDate=$(GIT_DATE)" ./cmd/shisui
 	mkdir -p $(GOBIN)
-	mv main $(GOBIN)/shisui
+	mv shisui $(GOBIN)/shisui
 	@echo "Done building."
 	@echo "Run \"$(GOBIN)/shisui\" to launch shisui."
